refactor: normalize file names to snake_case and fix typos
- Fix malformed badge_slot_provider..dart (double dot)
- Rename PascalCase/lowercase files to snake_case (BadgeScanProvider,
  imageprovider, getitlocator, homescreentabs, speedial, transitiontab,
  vectorview, badgeeffectabstract)
- Fix typo about_us_daialog -> about_us_dialog
- Rename homescreen -> home_screen, saved_clipart -> saved_clipart_screen
- Move stray widgets (badge_scan_settings_widget, special_text_field) into view/widgets
- Update all import paths accordingly
