+++ lib/badge_animation/ani_equalizer.dart
