+++ lib/badge_effect/badge_effect_abstract.dart
