+++ lib/badgemagic_module/bluetooth/datagenerator.dart
@@ -4,7 +4,7 @@ import 'package:badgemagic/badgemagic_module/models/data.dart';
 import 'package:badgemagic/badgemagic_module/utils/data_to_bytearray_converter.dart';
 import 'package:badgemagic/badgemagic_module/utils/file_helper.dart';
 import 'package:badgemagic/providers/badge_message_provider.dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:universal_ble/universal_ble.dart';
 import 'package:get_it/get_it.dart';
