+++ lib/badgemagic_module/bluetooth/scan_state.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'package:badgemagic/badgemagic_module/bluetooth/connect_state.dart';
 import 'package:badgemagic/badgemagic_module/bluetooth/datagenerator.dart';
-import 'package:badgemagic/providers/BadgeScanProvider.dart';
+import 'package:badgemagic/providers/badge_scan_provider.dart';
 import 'package:get_it/get_it.dart';
 import '../../services/localization_service.dart';
 import '../../view/widgets/ble_progress_dialog.dart';
