+++ lib/badgemagic_module/utils/byte_array_utils.dart
