+++ lib/badgemagic_module/utils/converters.dart
@@ -6,7 +6,7 @@ import 'package:badgemagic/badgemagic_module/utils/data_to_bytearray_converter.d
 import 'package:badgemagic/badgemagic_module/utils/file_helper.dart';
 import 'package:badgemagic/badgemagic_module/utils/image_utils.dart';
 import 'package:badgemagic/providers/font_provider.dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
