+++ lib/badgemagic_module/utils/file_helper.dart
@@ -6,7 +6,7 @@ import 'package:badgemagic/badgemagic_module/models/data.dart';
 import 'package:badgemagic/badgemagic_module/utils/byte_array_utils.dart';
 import 'package:badgemagic/badgemagic_module/utils/image_utils.dart';
 import 'package:badgemagic/badgemagic_module/utils/toast_utils.dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
