+++ lib/main.dart
@@ -2,15 +2,15 @@ import 'dart:math' as math;
 
 import 'package:badgemagic/providers/animation_badge_provider.dart';
 import 'package:badgemagic/providers/font_provider.dart';
-import 'package:badgemagic/providers/BadgeScanProvider.dart';
-import 'package:badgemagic/providers/getitlocator.dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/badge_scan_provider.dart';
+import 'package:badgemagic/providers/service_locator.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:badgemagic/providers/speed_dial_provider.dart';
 import 'package:badgemagic/view/about_us_screen.dart';
 import 'package:badgemagic/view/draw_badge_screen.dart';
-import 'package:badgemagic/view/homescreen.dart';
+import 'package:badgemagic/view/home_screen.dart';
 import 'package:badgemagic/view/save_badge_screen.dart';
-import 'package:badgemagic/view/saved_clipart.dart';
+import 'package:badgemagic/view/saved_clipart_screen.dart';
 import 'package:badgemagic/view/settings_screen.dart';
 import 'package:badgemagic/view/widgets/ble_progress_dialog_controller.dart';
 import 'package:flutter/material.dart';
