+++ lib/providers/badge_message_provider.dart
@@ -9,8 +9,8 @@ import 'package:badgemagic/badgemagic_module/models/data.dart';
 import 'package:badgemagic/badgemagic_module/models/messages.dart';
 import 'package:badgemagic/badgemagic_module/models/mode.dart';
 import 'package:badgemagic/badgemagic_module/models/speed.dart';
-import 'package:badgemagic/providers/BadgeScanProvider.dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/badge_scan_provider.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:badgemagic/services/localization_service.dart';
 import 'package:flutter/material.dart';
 import 'package:badgemagic/utils/custom_transfers/transfers.dart';
