+++ lib/providers/badge_scan_provider.dart
