+++ lib/providers/badge_slot_provider.dart
