+++ lib/providers/inline_image_provider.dart
