+++ lib/view/home_screen.dart
@@ -15,19 +15,19 @@ import 'package:badgemagic/providers/animation_badge_provider.dart';
 import 'package:badgemagic/providers/badge_message_provider.dart'
     hide modeValueMap, speedMap;
 import 'package:badgemagic/providers/font_provider.dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:badgemagic/providers/saved_badge_provider.dart';
 import 'package:badgemagic/providers/speed_dial_provider.dart';
 import 'package:badgemagic/services/localization_service.dart';
-import 'package:badgemagic/view/special_text_field.dart';
+import 'package:badgemagic/view/widgets/special_text_field.dart';
 import 'package:badgemagic/view/widgets/ble_progress_dialog.dart';
 import 'package:badgemagic/view/widgets/ble_progress_dialog_controller.dart';
 import 'package:badgemagic/view/widgets/common_scaffold_widget.dart';
-import 'package:badgemagic/view/widgets/homescreentabs.dart';
+import 'package:badgemagic/view/widgets/home_screen_tabs.dart';
 import 'package:badgemagic/view/widgets/save_badge_dialog.dart';
-import 'package:badgemagic/view/widgets/speedial.dart';
-import 'package:badgemagic/view/widgets/transitiontab.dart';
-import 'package:badgemagic/view/widgets/vectorview.dart';
+import 'package:badgemagic/view/widgets/speed_dial.dart';
+import 'package:badgemagic/view/widgets/transition_tab.dart';
+import 'package:badgemagic/view/widgets/vector_view.dart';
 import 'package:badgemagic/virtualbadge/view/animated_badge.dart';
 import 'package:extended_text_field/extended_text_field.dart';
 import 'package:flutter/material.dart';
