+++ lib/view/saved_clipart_screen.dart
@@ -1,7 +1,7 @@
 import 'package:badgemagic/badgemagic_module/utils/byte_array_utils.dart';
 import 'package:badgemagic/badgemagic_module/utils/file_helper.dart';
 import 'package:badgemagic/constants.dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:badgemagic/services/localization_service.dart';
 import 'package:badgemagic/view/widgets/clipart_list_view.dart';
 import 'package:badgemagic/view/widgets/common_scaffold_widget.dart';
