+++ lib/view/settings_screen.dart
@@ -1,5 +1,5 @@
 import 'package:badgemagic/constants.dart';
-import 'package:badgemagic/providers/BadgeScanProvider.dart';
+import 'package:badgemagic/providers/badge_scan_provider.dart';
 import 'package:badgemagic/view/widgets/common_scaffold_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
