+++ lib/view/widgets/save_badge_card.dart
@@ -7,11 +7,11 @@ import 'package:badgemagic/constants.dart';
 import 'package:badgemagic/l10n/app_localizations.dart';
 import 'package:badgemagic/providers/animation_badge_provider.dart';
 import 'package:badgemagic/providers/badge_message_provider.dart';
-import 'package:badgemagic/providers/badge_slot_provider..dart';
-import 'package:badgemagic/providers/imageprovider.dart';
+import 'package:badgemagic/providers/badge_slot_provider.dart';
+import 'package:badgemagic/providers/inline_image_provider.dart';
 import 'package:badgemagic/providers/saved_badge_provider.dart';
 import 'package:badgemagic/services/localization_service.dart';
-import 'package:badgemagic/view/homescreen.dart';
+import 'package:badgemagic/view/home_screen.dart';
 import 'package:badgemagic/view/widgets/badge_delete_dialog.dart';
 import 'package:badgemagic/view/widgets/qr_share_dialog.dart';
 import 'package:flutter/material.dart';
